requirements.txt
Django==5.2

django-cors-headers==4.7.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
social-auth-app-django==5.4.2

django-ckeditor==6.7.1

pytz==2025.1

Pillow==11.1.0
mutagen==1.47.0

psycopg2-binary==2.9.10

argon2-cffi==23.1.0

stripe==11.5.0
braintree==4.29.0

pyJWT==2.10.1

boto3==1.37.0
django-storages==1.14.5

markdown==3.7
gunicorn==23.0.0
whitenoise==6.9.0

requests==2.32.3
